pyproject.toml
[project]
name = "gurupod"
version = "0.1.0"
description = "Decoding The Guru Podcast wiki generator"
authors = [
    { name = "pawrequest", email = "paw@pawrequest.com" }
]
dependencies = [
    'sqlmodel',
    "uvicorn[standard]",
    'fastapi',
    'pydantic',
    'python-dotenv',
    "praw",
    "requests",
    "beautifulsoup4",
    "python-dateutil",
    "aiohttp",
    "pytest-asyncio",
    "asyncpraw",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "httpx",
    "pytest",
    "coverage>=7.3.2",
]

[tool.hatch.metadata]
allow-direct-references = true
